requirements_dev.txt
numpy==1.12.1
pandas==0.20.1
pip==9.0.1
bumpversion==0.5.3
wheel==0.29.0
watchdog==0.8.3
flake8==3.3.0
tox==2.7.0
coverage==4.4.1
Sphinx==1.6.1
cryptography==1.8.1
PyYAML==3.12
pytest==3.0.7
psycopg2==2.7.1
pymssql==2.1.3
dedupe==1.6.13
testing.postgresql==1.3.0
Faker==0.7.12
tqdm==4.11.2
pytest-cov==2.5.1
unicodecsv==0.14.1